+++ app/client/src/components/editorComponents/CodeEditor/EditorConfig.ts
@@ -5,21 +5,7 @@ import type { AutocompleteDataType } from "utils/autocomplete/AutocompleteDataTy
 import type { EntityNavigationData } from "selectors/navigationSelectors";
 import type { ExpectedValueExample } from "utils/validation/common";
 
-import { find } from "lodash";
-
-export const editorSQLModes = {
-  // SQL only
-  SQL: "sql",
-  // SQL flavour + JS
-  SNOWFLAKE_WITH_BINDING: "snowflakesql-js",
-  ARANGO_WITH_BINDING: "arangosql-js",
-  REDIS_WITH_BINDING: "redissql-js",
-  POSTGRESQL_WITH_BINDING: "pgsql-js",
-  SQL_WITH_BINDING: "sql-js",
-  MYSQL_WITH_BINDING: "mysql-js",
-  MSSQL_WITH_BINDING: "mssql-js",
-  PLSQL_WITH_BINDING: "plsql-js",
-} as const;
+import { editorSQLModes } from "./sql/config";
 
 export const EditorModes = {
   TEXT: "text/plain",
@@ -34,77 +20,6 @@ export const EditorModes = {
 
 type ValueOf<T> = T[keyof T];
 export type TEditorModes = ValueOf<typeof EditorModes>;
-export type TEditorSqlModes = ValueOf<typeof editorSQLModes>;
-type SqlModeConfig = Record<
-  TEditorSqlModes,
-  {
-    mime: string;
-    mode: TEditorSqlModes;
-    // CodeMirror.multiplexingMode
-    isMultiplex: boolean;
-  }
->;
-
-// Mime available in sql mode https://github.com/codemirror/codemirror5/blob/9974ded36bf01746eb2a00926916fef834d3d0d0/mode/sql/sql.js#L290
-export const sqlModesConfig: SqlModeConfig = {
-  [editorSQLModes.SQL]: {
-    mime: "sql",
-    mode: editorSQLModes.SQL,
-    isMultiplex: false,
-  },
-  [editorSQLModes.SQL_WITH_BINDING]: {
-    mime: "text/x-sql",
-    mode: editorSQLModes.SQL_WITH_BINDING,
-    isMultiplex: true,
-  },
-  [editorSQLModes.MYSQL_WITH_BINDING]: {
-    mime: "text/x-mysql",
-    mode: editorSQLModes.MYSQL_WITH_BINDING,
-    isMultiplex: true,
-  },
-  [editorSQLModes.MSSQL_WITH_BINDING]: {
-    mime: "text/x-mssql",
-    mode: editorSQLModes.MSSQL_WITH_BINDING,
-    isMultiplex: true,
-  },
-  [editorSQLModes.PLSQL_WITH_BINDING]: {
-    mime: "text/x-plsql",
-    mode: editorSQLModes.PLSQL_WITH_BINDING,
-    isMultiplex: true,
-  },
-  [editorSQLModes.POSTGRESQL_WITH_BINDING]: {
-    mime: "text/x-pgsql",
-    mode: editorSQLModes.POSTGRESQL_WITH_BINDING,
-    isMultiplex: true,
-  },
-  // Custom mimes
-  [editorSQLModes.SNOWFLAKE_WITH_BINDING]: {
-    mime: "text/x-snowflakesql",
-    mode: editorSQLModes.SNOWFLAKE_WITH_BINDING,
-    isMultiplex: true,
-  },
-  [editorSQLModes.ARANGO_WITH_BINDING]: {
-    mime: "text/x-arangosql",
-    mode: editorSQLModes.ARANGO_WITH_BINDING,
-    isMultiplex: true,
-  },
-  [editorSQLModes.REDIS_WITH_BINDING]: {
-    mime: "text/x-redis",
-    mode: editorSQLModes.REDIS_WITH_BINDING,
-    isMultiplex: true,
-  },
-};
-
-export const pluginNameToMIME: Record<string, ValueOf<typeof EditorModes>> = {
-  PostgreSQL: EditorModes.POSTGRESQL_WITH_BINDING,
-  MySQL: EditorModes.MYSQL_WITH_BINDING,
-  "Microsoft SQL Server": EditorModes.MSSQL_WITH_BINDING,
-  Oracle: EditorModes.PLSQL_WITH_BINDING,
-  Redshift: EditorModes.PLSQL_WITH_BINDING,
-  Snowflake: EditorModes.SNOWFLAKE_WITH_BINDING,
-  ArangoDB: EditorModes.ARANGO_WITH_BINDING,
-  Redis: EditorModes.REDIS_WITH_BINDING,
-};
 
 export enum EditorTheme {
   LIGHT = "LIGHT",
@@ -123,7 +38,7 @@ export enum EditorSize {
 
 export type EditorConfig = {
   theme: EditorTheme;
-  mode: ValueOf<typeof EditorModes>;
+  mode: TEditorModes;
   tabBehaviour: TabBehaviour;
   size: EditorSize;
   hinting?: Array<HintHelper>;
@@ -219,16 +134,3 @@ export const INDENTATION_CHARACTERS = {
 export const isNavKey = (key: any): key is AUTOCOMPLETE_NAVIGATION => {
   return AUTOCOMPLETE_NAVIGATION.hasOwnProperty(key);
 };
-
-export function getSqlEditorModeFromPluginName(name: string) {
-  return pluginNameToMIME[name] ?? EditorModes.SQL_WITH_BINDING;
-}
-
-export function getSqlMimeFromMode(mode: ValueOf<typeof editorSQLModes>) {
-  const modeConfig = find(sqlModesConfig, { mode });
-  return modeConfig?.mime ?? "text/x-sql";
-}
-
-export function isSqlMode(mode: TEditorModes) {
-  return !!Object.keys(sqlModesConfig).includes(mode);
-}
