+++ app/client/src/components/editorComponents/CodeEditor/index.tsx
@@ -37,7 +37,7 @@ import type {
 import { ENTITY_TYPE } from "entities/DataTree/dataTreeFactory";
 import { Skin } from "constants/DefaultTheme";
 import AnalyticsUtil from "utils/AnalyticsUtil";
-import "components/editorComponents/CodeEditor/customMimes";
+import "components/editorComponents/CodeEditor/sql/customMimes";
 import "components/editorComponents/CodeEditor/modes";
 import type {
   CodeEditorBorder,
