+++ app/client/src/components/editorComponents/CodeEditor/sql/customMimes/index.ts
