+++ app/client/src/components/editorComponents/CodeEditor/sql/customMimes/utils.ts
